requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.28.84
botocore==1.31.84
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
daphne==4.0.0
distro==1.8.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.4
msgpack==1.0.7
multidict==6.0.4
openai==1.3.8
Pillow==10.1.0
prompt-toolkit==3.0.41
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==7.5.0
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
yarl==1.9.2
zope.interface==6.1
